requirements.txt
Flask==3.0.3
gunicorn==22.0.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
python-dotenv==1.0.1

WazeRouteCalculator==0.16
requests==2.32.3
